requirements.txt
Flask==0.12.1
mysql-connector-python==8.0.6
azure-storage-blob==1.1.0
python-dateutil